requirements.txt
nose
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
decorator==4.0.10
docker-py==1.8.1
gnureadline==6.3.3
ipaddress==1.0.16
pathlib2==2.1.0
pexpect==4.1.0
pickleshare==0.7.2
ptyprocess==0.5.1
requests==2.10.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.1
websocket-client==0.37.0
